connect liked journals
