+++ frontend/src/screens/Explore.js
@@ -224,6 +224,7 @@ export default function Explore() {
                     searchContent={searchContent}
                     handleSearchChange={handleSearchChange}
                     handleSearch={handleSearch}
+                    isSearchEnabled={true}
                 />
                 <div className={classes.explore_bg} />
                 <ExploreTabs handleTab={handleTab} />
