+++ frontend/src/screens/Me.js
@@ -253,6 +253,7 @@ export default function Me() {
                     searchContent={searchContent}
                     handleSearchChange={handleSearchChange}
                     handleSearch={handleSearch}
+                    isSearchEnabled={true}
                 />
 
                 <div className={classes.my_journals_bg} />
